feat(register): improve ui
